+++ type.go
@@ -3,5 +3,5 @@ package tsdbParser
 type Metric struct {
 	MetricName 	string `json:"metric"`
 	Tags   		map[string]string `json:"tags"`
-	Dps 		map[int64]int `json:"dps"`
+	Dps 		map[int64]float64 `json:"dps"`
 }
